refactor(backstop): moves vars file and updates scenarios
