+++ _ci/backstop/scenarios/post/post_anchor_active_aside.js
@@ -1,12 +1,12 @@
-const { URL } = require('../../vars');
+const { URL } = require('../vars');
 const POST_URL = `${URL}/post/chicken-caesar-salad-an-overrated-concoction-of-lettuce-and-delusion#step-2-caesar-dressing-the-holy-grail-of-salad-sauces`;
-const { aboveSmallViewports } = require('../../vars.js');
+const { aboveSmallViewports } = require('../vars.js');
 
 module.exports = {
   label: 'Post - Anchor Active Aside',
   url: POST_URL,
   viewports: aboveSmallViewports,
-  onReadyScript: 'puppet/postScrollToTop.js',
+  onReadyScript: 'playwright/postScrollToTop.js',
   delay: 500,
   scrollToSelector: 'div.page-navigation-list a.active',
   postInteractionWait: 500,
