+++ _ci/backstop/scenarios/posts/posts_hover_post_1.js
@@ -1,20 +1,13 @@
-const { URL } = require('../../vars');
+const { URL } = require('../vars');
 const POSTS_URL = `${URL}/posts`;
 
 module.exports = {
   label: 'Posts - Hover First Post',
   url: POSTS_URL,
-  readyEvent: '',
-  readySelector: '',
-  onReadyScript: 'puppet/onReadyPosts.js',
-  delay: 800,
-  hideSelectors: [],
-  removeSelectors: [],
+  onReadyScript: 'playwright/onReadyPosts.js',
   hoverSelector: '.posts-grid li:nth-child(1) .post-link',
-  postInteractionWait: 1000,
+  postInteractionWait: 2500,
   selectors: ['document'],
-  selectorExpansion: true,
-  expect: 0,
   misMatchThreshold: 0.2,
   requireSameDimensions: true,
 };
