+++ _ci/backstop/scenarios/posts/posts_hover_post_7.js
@@ -1,20 +1,15 @@
-const { URL } = require('../../vars');
+const { URL } = require('../vars');
 const POSTS_URL = `${URL}/posts`;
 
 module.exports = {
   label: 'Posts - Hover Seventh Post',
   url: POSTS_URL,
-  readyEvent: '',
-  readySelector: '',
-  onReadyScript: 'puppet/onReadyPosts.js',
-  delay: 800,
-  hideSelectors: [],
-  removeSelectors: [],
+  onReadyScript: 'playwright/onReadyPosts.js',
+  delay: 1000,
   hoverSelector: '.posts-grid li:nth-child(7) .post-link',
-  postInteractionWait: 1000,
+  postInteractionWait: 2500,
   selectors: ['document'],
-  selectorExpansion: true,
-  expect: 0,
+  selectorExpansion: false,
   misMatchThreshold: 0.2,
   requireSameDimensions: true,
 };
